Layout graph with text, and link arrows.
New csv file from pathfinder community, has some duplicates that need
handling.

Link stroke and arrow are slightly opaque.

Increase zoom to 8.

Trying some force parameters to speed up simulation times.
More work needs to be done there.

Node creation needs an initial x, y as these were throwing NaN instead.
This may have been because some values were set and not all.
Setting everything to be at 0,0 is an issue for the simulation as the
charge factor flings things too far and having NaN, -Infinity, or
+Infinity isn't good for the simulation. Instead they are initialized as
a grid, but once a decent layout is achieved the current x,y values need
to be captured and used to seed the simulation.

Mythic feats have the same name, but a different category. To avoid
these being duplicates the name is changed to include a (Mythic) suffix.

Draw order is important for svg, what is drawn first is lower in the
z-order and hence will be over-drawn with higher order values. So links
need to get drawn first so that the node will draw over the top of the
line end points. The proper way to fix that would be to adjust the line
length to be shorter on each end by the node's radius, but that is more
maths than necessary. The current hack just moves the arrow along the
x-axis so that it pokes out near the node circumference.

Nodes are also an svg g element, that contains the circle and the text.
This means that translate() is needed to move them around instead of
adjusting the circles point of origin.