Add layout status notification. Add prerequisites x,y.
Indicate the alpha value in the layout notification as an indicator of
how much longer the simulation will run for.

Update scaling to be 1/32 to 32 zoom.

Add the x,y for prerequisites.
Debug logs for NaN x,y values are done addNode time.
Duplicates are logged to console and then ignored, as the split between
feats and prerequisites made it more likely duplicates were created.

Split load link and building node list into separate functions.

Nodes that don't have an x,y get initialized to a position on a grid.